refactor: Update geotag page to use last coordinates instead of initial route point
