+++ lib/screens/_starting.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:pcic_mobile_app/screens/dashboard/_home.dart';
-import 'package:pcic_mobile_app/utils/_app_session.dart';
-import 'package:pcic_mobile_app/utils/authentication/_login.dart';
-import 'package:pcic_mobile_app/utils/authentication/_signup.dart';
+import 'package:pcic_mobile_app/screens/home/_home.dart';
+import 'package:pcic_mobile_app/utils/agent/_session.dart';
+import 'package:pcic_mobile_app/utils/agent/_login.dart';
+import 'package:pcic_mobile_app/utils/agent/_signup.dart';
 
 class StartingPage extends StatelessWidget {
   const StartingPage({super.key});
