+++ lib/screens/geotag/_geotag.dart
@@ -5,8 +5,8 @@ import 'package:external_path/external_path.dart';
 import 'package:flutter/material.dart';
 import 'package:gpx/gpx.dart';
 import 'package:latlong2/latlong.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/forms_components/_pcic_form.dart';
-import 'package:pcic_mobile_app/utils/_app_gpx.dart';
+import 'package:pcic_mobile_app/screens/pcic_form/_pcic_form.dart';
+import 'package:pcic_mobile_app/utils/app/_gpx.dart';
 import 'package:pcic_mobile_app/utils/controls/_control_task.dart';
 import 'package:pcic_mobile_app/utils/controls/_location_service.dart';
 import 'package:pcic_mobile_app/utils/controls/_map_service.dart';
