+++ lib/screens/home/_home.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:pcic_mobile_app/screens/_splash.dart';
-import 'package:pcic_mobile_app/screens/dashboard/_message.dart';
-import 'package:pcic_mobile_app/screens/dashboard/_task.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/home_components/_home_header.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/home_components/_profile_container.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/home_components/_recent_task_container.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/home_components/_search_button.dart';
+import 'package:pcic_mobile_app/screens/messages/_view.dart';
+import 'package:pcic_mobile_app/screens/tasks/_task.dart';
+import 'package:pcic_mobile_app/screens/home/_home_header.dart';
+import 'package:pcic_mobile_app/screens/home/_profile_container.dart';
+import 'package:pcic_mobile_app/screens/home/_recent_task_container.dart';
+import 'package:pcic_mobile_app/screens/home/_search_button.dart';
 import 'package:pcic_mobile_app/utils/controls/_control_task.dart';
 import 'package:shared_preferences/shared_preferences.dart';
