+++ lib/screens/messages/_view.dart
@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:pcic_mobile_app/utils/controls/_filter_message.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/_chat.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/_message_items.dart';
+import 'package:pcic_mobile_app/screens/messages/_filter.dart';
+import 'package:pcic_mobile_app/screens/messages/_detail.dart';
+import 'package:pcic_mobile_app/screens/messages/_items.dart';
 import 'package:provider/provider.dart';
 
 class MessagesPage extends StatefulWidget {
   const MessagesPage({super.key});
 
   @override
-  _MessagesPageState createState() => _MessagesPageState();
+  MessagesPageState createState() => MessagesPageState();
 }
 
-class _MessagesPageState extends State<MessagesPage> {
+class MessagesPageState extends State<MessagesPage> {
   final List<Map<String, dynamic>> messages = [
     {
       'name': 'Joemar Cardiño',
