+++ lib/screens/pcic_form/_pcic_form.dart
@@ -6,10 +6,10 @@ import 'package:intl/intl.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:open_file/open_file.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/forms_components/_date_control.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/forms_components/_success.dart';
-import 'package:pcic_mobile_app/screens/dashboard/views/tasks_components/_signature_section.dart';
-import 'package:pcic_mobile_app/utils/controls/_control_actual_seeds.dart';
+import 'package:pcic_mobile_app/screens/pcic_form/_date_control.dart';
+import 'package:pcic_mobile_app/screens/pcic_form/_success.dart';
+import 'package:pcic_mobile_app/screens/tasks/_signature_section.dart';
+import 'package:pcic_mobile_app/utils/seeds/_dropdown.dart';
 import 'package:pcic_mobile_app/utils/controls/_control_task.dart';
 import 'package:pcic_mobile_app/utils/controls/_map_service.dart';
 import 'package:permission_handler/permission_handler.dart';
