+++ lib/utils/agent/_logout_view.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
-import 'package:pcic_mobile_app/utils/_app_colors.dart';
-import 'package:pcic_mobile_app/utils/authentication/_login.dart';
+import 'package:pcic_mobile_app/utils/app/_colors.dart';
+import 'package:pcic_mobile_app/utils/agent/_login.dart';
 
 class LogoutSuccessPage extends StatelessWidget {
   const LogoutSuccessPage({super.key});
