+++ lib/utils/agent/_session.dart
